requirements.txt
asgiref==3.5.2
boto==2.49.0
boto3==1.24.42
botocore==1.27.42
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
dj-database-url==0.5.0
Django==4.0.6
django-address==0.2.8
django-auth-adfs==1.10.0
django-concurrency==2.3
django-crispy-forms==1.14.0
django-easy-select2==1.5.7
django-filter==22.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-shortcuts==1.6
django-storages==1.12.3
djangorestframework==3.13.1
easy-date==0.1.4
idna==3.3
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.20.0
pytz==2021.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzdata==2022.1
urllib3==1.26.6
whitenoise==5.3.0
zipp==3.5.0
